requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
bcrypt==3.2.0
black==21.9b0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
ecdsa==0.17.0
fastapi==0.68.1
h11==0.12.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.3.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-jose==3.3.0
python-multipart==0.0.5
regex==2021.8.28
rsa==4.7.2
six==1.16.0
starlette==0.14.2
tomli==1.2.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
